requirements.txt
pylint
flake8
pandas
rdflib
python-dotenv